+++ R/scorecard.R
@@ -35,7 +35,7 @@ ab = function(points0=600, odds0=1/19, pdo=50) {
 #' @param odds0 Target odds, default 1/19. Odds = p/(1-p).
 #' @param pdo Points to Double the Odds, default 50.
 #' @param basepoints_eq0 Logical, Defaults to FALSE. If it is TRUE, the basepoints will equally distribute to each variable.
-#' @return A scorecard data frame
+#' @return A list of scorecard data frames
 #'
 #' @seealso \code{\link{scorecard2}} \code{\link{scorecard_ply}}
 #'
@@ -71,11 +71,7 @@ ab = function(points0=600, odds0=1/19, pdo=50) {
 #' # Example I # creat a scorecard
 #' card = scorecard(bins, m)
 #' card2 = scorecard2(bins=bins, dt=germancredit, y='creditability',
-#'   x=c("status.of.existing.checking.account", "duration.in.month", "credit.history",
-#'    "purpose", "credit.amount", "savings.account.and.bonds",
-#'    "present.employment.since", "installment.rate.in.percentage.of.disposable.income",
-#'    "personal.status.and.sex", "other.debtors.or.guarantors", "property",
-#'    "age.in.years", "other.installment.plans", "housing"))
+#'   x=sub('_woe', '', names(coef(m))[-1]))
 #'
 #' # credit score
 #' # Example I # only total score
@@ -126,19 +122,21 @@ scorecard = function(bins, model, points0=600, odds0=1/19, pdo=50, basepoints_eq
 
 #' Creating a Scorecard
 #'
-#' \code{scorecard2} creates a scorecard based on the results from \code{woebin}. It has the same function with \code{scorecard}, but without model object input.
+#' \code{scorecard2} creates a scorecard based on the results from \code{woebin}. It has the same function of \code{scorecard}, but without model object input and provided adjustment for oversampling.
 #'
+#' @param bins Binning information generated from \code{woebin} function.
 #' @param dt A data frame with both x (predictor/feature) and y (response/label) variables.
 #' @param y Name of y variable.
 #' @param x Name of x variables. If it is NULL, then all variables in bins are used. Defaults to NULL.
-#' @param bins Binning information generated from \code{woebin} function.
+#' @param badprob_pop Bad probability of population. Accepted range: 0-1,  default to NULL. If it is not NULL, the model will adjust for oversampling.
 #' @param points0 Target points, default 600.
 #' @param odds0 Target odds, default 1/19. Odds = p/(1-p).
 #' @param pdo Points to Double the Odds, default 50.
-#' @param basepoints_eq0 Logical, Defaults to FALSE. If it is TRUE, the basepoints will equally distribute to each variable.
+#' @param basepoints_eq0 Logical, defaults to FALSE. If it is TRUE, the basepoints will equally distribute to each variable.
+#' @param return_prob Logical, defaults to FALSE. If it is TRUE, the predict probability will also return.
 #' @param positive Value of positive class, default "bad|1".
 #' @param ... Additional parameters.
-#' @return A scorecard data frame
+#' @return A list of scorecard data frames
 #'
 #' @seealso \code{\link{scorecard}} \code{\link{scorecard_ply}}
 #'
@@ -174,11 +172,7 @@ scorecard = function(bins, model, points0=600, odds0=1/19, pdo=50, basepoints_eq
 #' # Example I # creat a scorecard
 #' card = scorecard(bins, m)
 #' card2 = scorecard2(bins=bins, dt=germancredit, y='creditability',
-#'   x=c("status.of.existing.checking.account", "duration.in.month", "credit.history",
-#'    "purpose", "credit.amount", "savings.account.and.bonds",
-#'    "present.employment.since", "installment.rate.in.percentage.of.disposable.income",
-#'    "personal.status.and.sex", "other.debtors.or.guarantors", "property",
-#'    "age.in.years", "other.installment.plans", "housing"))
+#'   x= sub('_woe', '', names(coef(m))[-1]))
 #'
 #' # credit score
 #' # Example I # only total score
@@ -188,9 +182,10 @@ scorecard = function(bins, model, points0=600, odds0=1/19, pdo=50, basepoints_eq
 #' score2 = scorecard_ply(germancredit, card, only_total_score = F)
 #' }
 #' @import data.table
+#' @importFrom stats predict
 #' @export
-scorecard2 = function(bins, dt, y, x=NULL, points0=600, odds0=1/19, pdo=50, basepoints_eq0=FALSE, positive='bad|1', ...) {
-  variable = NULL
+scorecard2 = function(bins, dt, y, x=NULL, badprob_pop = NULL, points0=600, odds0=1/19, pdo=50, basepoints_eq0=FALSE, return_prob = FALSE, positive='bad|1', ...) {
+  variable = wgts = NULL
 
   dt = setDT(copy(dt))
   # bins # if (is.list(bins)) rbindlist(bins)
@@ -207,12 +202,31 @@ scorecard2 = function(bins, dt, y, x=NULL, points0=600, odds0=1/19, pdo=50, base
   # dt to woe values
   dt_woe = do.call(woebin_ply, args = c(list(dt=dt, bins=bins, print_info=FALSE), list(...)))
 
-  # model
-  model = glm(
-    as.formula(paste(y, "~ .")), family = binomial(),
-    data = dt_woe[,c(paste0(x,"_woe"),y), with=F])
 
-  return(scorecard(bins = bins, model = model, points0 = points0, odds0 = odds0, pdo = pdo, basepoints_eq0 = basepoints_eq0))
+  # model
+  if (!is.null(badprob_pop) && badprob_pop > 0 && badprob_pop < 1) {
+    p1 = badprob_pop # bad probability in population
+    r1 = dt[, table(get(y))/.N][['1']] # bad probability in sample dataset
+    dt_woe = dt_woe[grepl(positive, get(y)), wgts := p1/r1
+                  ][is.na(wgts), wgts := (1-p1)/(1-r1)
+                  ][,c(paste0(x,"_woe"), "wgts", y), with=F]
+
+    fmla = as.formula(sprintf('%s ~ %s', y, paste(paste0(x,"_woe"), collapse=" + ")))
+    model = glm(fmla, family = 'quasibinomial', weights = wgts, data = dt_woe)
+  } else {
+
+    model = glm(
+      as.formula(paste(y, "~ .")), family = 'binomial',
+      data = dt_woe[,c(paste0(x,"_woe"),y), with=F])
+  }
+
+
+  card = scorecard(bins = bins, model = model, points0 = points0, odds0 = odds0, pdo = pdo, basepoints_eq0 = basepoints_eq0)
+  if (return_prob) rt = list(
+    card = card,
+    prob = predict(model, dt_woe, type='response')
+  ) else rt = card
+  return(rt)
 }
 
 #' Score Transformation
@@ -262,11 +276,7 @@ scorecard2 = function(bins, dt, y, x=NULL, points0=600, odds0=1/19, pdo=50, base
 #' # Example I # creat a scorecard
 #' card = scorecard(bins, m)
 #' card2 = scorecard2(bins=bins, dt=germancredit, y='creditability',
-#'   x=c("status.of.existing.checking.account", "duration.in.month", "credit.history",
-#'    "purpose", "credit.amount", "savings.account.and.bonds",
-#'    "present.employment.since", "installment.rate.in.percentage.of.disposable.income",
-#'    "personal.status.and.sex", "other.debtors.or.guarantors", "property",
-#'    "age.in.years", "other.installment.plans", "housing"))
+#'   x=sub('_woe', '', names(coef(m))[-1]))
 #'
 #' # credit score
 #' # Example I # only total score