+++ R/woebin.R
@@ -485,7 +485,7 @@ woebin2_chimerge = function(
     e_lag = a_lag_rowsum*a_colsum/a_sum
   )][, .(chisq=sum((a-e)^2/e + (a_lag-e_lag)^2/e_lag)), by='bin']
 
-  return(merge(initial_binning[,count:=good+bad], chisq_df, all.x = TRUE))
+  return(merge(initial_binning[,count:=good+bad], chisq_df, all.x = TRUE, sort = FALSE))
   }
 
   # dtm_rows
@@ -518,7 +518,7 @@ woebin2_chimerge = function(
     } else if (bin_nrow > bin_num_limit) {
       rm_brkp = binning_chisq[, merge_tolead := FALSE][order(chisq, count)][1,]
 
-    }
+    } else break
 
     # groupby brkp
     shift_type = ifelse(rm_brkp[1,merge_tolead], 'lead', 'lag')
@@ -819,11 +819,9 @@ bins_to_breaks = function(bins, dt, to_string=FALSE, save_name=NULL) {
 #' @importFrom parallel detectCores makeCluster stopCluster
 #' @export
 woebin = function(
-  dt, y, x=NULL,
-  var_skip=NULL, breaks_list=NULL, special_values=NULL,
+  dt, y, x=NULL, var_skip=NULL, breaks_list=NULL, special_values=NULL,
   stop_limit=0.1, count_distr_limit=0.05, bin_num_limit=8,
-  positive="bad|1", no_cores=NULL, print_step=0L,
-  method="tree", save_breaks_list=NULL,
+  positive="bad|1", no_cores=NULL, print_step=0L, method="tree", save_breaks_list=NULL,
   ignore_const_cols=TRUE, ignore_datetime_cols=TRUE, check_cate_num=TRUE, replace_blank_inf=TRUE, ...) {
   # start time
   start_time = proc.time()
@@ -831,20 +829,28 @@ woebin = function(
   # global variable
   i = NULL
   # arguments ------
+  kwargs = list(...)
+  # method
+  method = match.arg(method, c("tree", "chimerge", 'freq', 'width'))
+  if (!(method %in% c("tree", "chimerge", 'freq', 'width'))) {
+    warning("Incorrect inputs; method should be tree or chimerge. Parameter was set to default (tree).")
+    method = "tree"
+  }
+  if (is.null(y) & !(method %in% c('freq', 'width'))) method = 'freq'
   # print_info
-  print_info = list(...)[['print_info']]
+  print_info = kwargs[['print_info']]
   if (is.null(print_info)) print_info = TRUE
   # init_count_distr
-  min_perc_fine_bin = list(...)[['init_count_distr']]
-  init_count_distr = list(...)[['init_count_distr']]
+  min_perc_fine_bin = kwargs[['init_count_distr']]
+  init_count_distr = kwargs[['init_count_distr']]
   if (is.null(init_count_distr)) {
     init_count_distr <- ifelse(!is.null(min_perc_fine_bin), min_perc_fine_bin, 0.02)
   }
   # count_distr_limit
-  min_perc_coarse_bin = list(...)[['min_perc_coarse_bin']]
+  min_perc_coarse_bin = kwargs[['min_perc_coarse_bin']]
   if (!is.null(min_perc_coarse_bin)) count_distr_limit = min_perc_coarse_bin
   # bin_num_limit
-  max_num_bin = list(...)[['max_num_bin']]
+  max_num_bin = kwargs[['max_num_bin']]
   if (!is.null(max_num_bin)) bin_num_limit = max_num_bin
 
 
@@ -865,7 +871,7 @@ woebin = function(
   # replace black with na
   if (replace_blank_inf) dt = rep_blank_na(dt)
   # x variable names
-  xs = x_variable(dt, y, x, var_skip)
+  xs = x_variable(dt, y, x, var_skip, method)
   xs_len = length(xs)
   # print_step
   print_step = check_print_step(print_step)
@@ -894,12 +900,8 @@ woebin = function(
     bin_num_limit = 8
   }
 
-  # method
-  if (!(method %in% c("tree", "chimerge", 'freq', 'width'))) {
-    warning("Incorrect inputs; method should be tree or chimerge. Parameter was set to default (tree).")
-    method = "tree"
-  }
-  if (is.null(y) & !(method %in% c('freq', 'width'))) method = 'freq'
+
+
 
   # binning ------
   # loop on xs
@@ -1033,12 +1035,13 @@ woepoints_ply1 = function(dtx, binx, x_i, woe_points) {
   return(dtx_suffix)
 }
 
-#' WOE Transformation
+#' WOE/BIN Transformation
 #'
-#' \code{woebin_ply} converts original input data into woe values based on the binning information generated from \code{woebin}.
+#' \code{woebin_ply} converts original values of input data into woe or bin based on the binning information generated from \code{woebin}.
 #'
 #' @param dt A data frame.
 #' @param bins Binning information generated from \code{woebin}.
+#' @param to Converting original values to woe or bin. Defaults to woe.
 #' @param no_cores Number of CPU cores for parallel computation. Defaults to 90 percent of total cpu cores.
 #' @param print_step A non-negative integer. Defaults to 1. If print_step>0, print variable names by each print_step-th iteration. If print_step=0 or no_cores>1, no message is print.
 #' @param replace_blank_inf Logical. Replace blank values with NA and infinite with -1. Defaults to TRUE. This argument should be the same with \code{woebin}'s.
@@ -1058,9 +1061,13 @@ woepoints_ply1 = function(dtx, binx, x_i, woe_points) {
 #' # binning for dt
 #' bins = woebin(dt, y = "creditability")
 #'
-#' # converting original value to woe
+#' # converting to woe
 #' dt_woe = woebin_ply(dt, bins=bins)
 #' str(dt_woe)
+#'
+#' # converting to bin
+#' dt_bin = woebin_ply(dt, bins=bins, to = 'bin')
+#' str(dt_bin)
 #'
 #' \donttest{
 #' # Example II
@@ -1075,24 +1082,24 @@ woepoints_ply1 = function(dtx, binx, x_i, woe_points) {
 #' bins_df = data.table::rbindlist(bins_germancredit)
 #' germancredit_woe = woebin_ply(germancredit, bins_df)
 #'
-#' # return value is bin but not woe
-#' germancredit_bin = woebin_ply(germancredit, bins_germancredit, value = 'bin')
 #' }
 #'
 #' @import data.table
 #' @export
 #'
-woebin_ply = function(dt, bins, no_cores=NULL, print_step=0L, replace_blank_inf=TRUE, ...) {
+woebin_ply = function(dt, bins, to='woe', no_cores=NULL, print_step=0L, replace_blank_inf=TRUE, ...) {
   # start time
   start_time = proc.time()
 
   # print info
-  print_info = list(...)[['print_info']]
+  kwargs = list(...)
+
+  print_info = kwargs[['print_info']]
   if (is.null(print_info)) print_info = TRUE
   if (print_info) cat('[INFO] converting into woe values ... \n')
-  # value
-  value = list(...)[['value']]
-  if (is.null(value) || !(value %in% c('woe', 'bin'))) value = 'woe'
+  # to
+  if (!is.null(kwargs[['value']])) to = kwargs[['value']]
+  if (is.null(to) || !(to %in% c('woe', 'bin'))) to = 'woe'
 
   # global variables or functions
   . = V1 = bin = variable = woe = i = databc_colomun_placeholder = NULL
@@ -1140,7 +1147,7 @@ woebin_ply = function(dt, bins, no_cores=NULL, print_step=0L, replace_blank_inf=
       binx = bins[variable==x_i]
       dtx = dt[, x_i, with=FALSE]
 
-      dat = cbind(dat, woepoints_ply1(dtx, binx, x_i, woe_points=value))
+      dat = cbind(dat, woepoints_ply1(dtx, binx, x_i, woe_points=to))
     }
   } else {
     # type_psock_fork = ifelse(Sys.info()["sysname"] == 'Windows', 'PSOCK', 'FORK')
@@ -1162,7 +1169,7 @@ woebin_ply = function(dt, bins, no_cores=NULL, print_step=0L, replace_blank_inf=
         binx = bins[variable==x_i]
         dtx = dt[, x_i, with=FALSE]
 
-        woepoints_ply1(dtx, binx, x_i, woe_points=value)
+        woepoints_ply1(dtx, binx, x_i, woe_points=to)
       }
     # finish
     stopCluster(cl)
@@ -1313,12 +1320,13 @@ woebin_plot = function(bins, x=NULL, title=NULL, show_iv = TRUE, line_value = 'b
   variable = NULL
   xs = x
 
+  kwargs = list(...)
   # line value
   if (!(line_value %in% c('badprob', 'woe'))) line_value = 'badprob'
   # line bar colors
-  line_color = list(...)[['line_color']]
+  line_color = kwargs[['line_color']]
   if (is.null(line_color)) line_color = 'blue'
-  bar_color = list(...)[['bar_color']]
+  bar_color = kwargs[['bar_color']]
 
   # bins # if (is.list(bins)) rbindlist(bins)
   if (inherits(bins, 'list') && all(sapply(bins, is.data.frame))) bins = rbindlist(bins)
@@ -1472,8 +1480,9 @@ woebin_adj = function(dt, y, bins, adj_all_var=TRUE, special_values=NULL, method
   # special_values
   special_values = check_special_values(special_values, xs_adj)
   # stop_limit
+  kwargs = list(...)
   stop_limit = 0.1
-  if ('stop_limit' %in% names(list(...))) stop_limit = list(...)[['stop_limit']]
+  if ('stop_limit' %in% names(kwargs)) stop_limit = kwargs[['stop_limit']]
     # print(stop_limit)
   stop_limit = check_stop_limit(stop_limit, xs_adj)